pyproject.toml
[tool.black]
line-length = 80

[tool.isort]
profile = "black"

[tool.ruff]
# Same as Black.
line-length = 80

[tool.ruff.lint]
select = ["D", "W505"]
ignore = ["D100", "D102", "D104", "D105", "D107", "D203", "D213", "D401", "D402"]

[tool.ruff.lint.pycodestyle]
max-doc-length = 100

[tool.ruff.lint.pydocstyle]
convention = "google"

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.setuptools]
packages = ["autoscholar", "quantmind"]

[project]
name = "quantmind"
version = "0.2.0"
description = "QuantMind - Intelligent Knowledge Extraction and Retrieval Framework"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "arxiv",
    "pyyaml",
    "ruff",
    "pytest",
    "numpy>=2.2.4",
    "networkx>=3.4.2",
    "scikit-learn>=1.6.1",
    "pyvis==0.3.1",
    "plotly>=6.0.1",
    "openai>=1.68.2",
    "camel-ai>=0.2.42",
    "pillow>=11.1.0",
    "llama-cloud-services>=0.6.12",
    "ipykernel>=6.29.5",
    "pydantic>=2.0.0",
    "pymupdf>=1.23.0",
]

[project.scripts]
quantmind = "quantmind_cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]
full = [
    "marker-pdf>=0.2.0",
    "beautifulsoup4>=4.12.0",
    "sentence-transformers>=2.2.0",
]
